+++ source/nyhal-linux-backend/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.29.6] - 2026-09-13
+
 ### Added
 
 - **Hot-swap documents into the live session**